+++ apps/server/.env
@@ -40,6 +40,9 @@ MOONSHOT_KEY=  # Your Moonshot API key
 # Lepton API Key
 LEPTON_KEY=  # Your Lepton API key (for accessing Lepton models)
 
+# JINA API KEY
+JINA_KEY=  # Your Jina API key
+
 # Local LLM Configuration (Ollama)
 OLLAMA_HOST=http://host.docker.internal:11434  # Hostname for Ollama (modify if needed for local LLM)
