+++ apps/server/src/libs/provider/tencent.ts
