+++ apps/server/src/libs/search/chatglm.ts
