+++ apps/server/src/service/rag/prompt.ts
